requirements.txt
streamlit
numpy
pillow
tensorflow
mediapipe
opencv-python-headless
gdown